feat(p2p): run periodic bootstrap discovery
